🎨 move the method pushArticleToRhy to B3ArticlePusher
